Refactor schema
