+++ samtranslator/schema/aws_serverless_statemachine.py
@@ -1,0 +1,142 @@
+from typing import Optional, Any, Dict, Union, List
+
+from typing_extensions import Literal
+
+from samtranslator.schema.common import PassThrough, BaseModel, SamIntrinsic
+
+
+class DeadLetterConfig(BaseModel):
+    Arn: Optional[PassThrough]
+    QueueLogicalId: Optional[str]
+    Type: Optional[Literal["SQS"]]
+
+
+class ScheduleEventProperties(BaseModel):
+    DeadLetterConfig: Optional[DeadLetterConfig]
+    Description: Optional[PassThrough]
+    Enabled: Optional[bool]
+    Input: Optional[PassThrough]
+    Name: Optional[PassThrough]
+    RetryPolicy: Optional[PassThrough]
+    Schedule: Optional[PassThrough]
+    State: Optional[PassThrough]
+
+
+class ScheduleEvent(BaseModel):
+    Type: Literal["Schedule"]
+    Properties: ScheduleEventProperties
+
+
+class ScheduleV2EventProperties(BaseModel):
+    DeadLetterConfig: Optional[DeadLetterConfig]
+    Description: Optional[PassThrough]
+    EndDate: Optional[PassThrough]
+    FlexibleTimeWindow: Optional[PassThrough]
+    GroupName: Optional[PassThrough]
+    Input: Optional[PassThrough]
+    KmsKeyArn: Optional[PassThrough]
+    Name: Optional[PassThrough]
+    PermissionsBoundary: Optional[PassThrough]
+    RetryPolicy: Optional[PassThrough]
+    RoleArn: Optional[PassThrough]  # TODO: Add to docs
+    ScheduleExpression: Optional[PassThrough]
+    ScheduleExpressionTimezone: Optional[PassThrough]
+    StartDate: Optional[PassThrough]
+    State: Optional[PassThrough]
+
+
+class ScheduleV2Event(BaseModel):
+    Type: Literal["ScheduleV2"]
+    Properties: ScheduleV2EventProperties
+
+
+class ResourcePolicy(BaseModel):
+    AwsAccountBlacklist: Optional[List[Union[str, Dict[str, Any]]]]
+    AwsAccountWhitelist: Optional[List[Union[str, Dict[str, Any]]]]
+    CustomStatements: Optional[List[Union[str, Dict[str, Any]]]]
+    IntrinsicVpcBlacklist: Optional[List[Union[str, Dict[str, Any]]]]
+    IntrinsicVpcWhitelist: Optional[List[Union[str, Dict[str, Any]]]]
+    IntrinsicVpceBlacklist: Optional[List[Union[str, Dict[str, Any]]]]
+    IntrinsicVpceWhitelist: Optional[List[Union[str, Dict[str, Any]]]]
+    IpRangeBlacklist: Optional[List[Union[str, Dict[str, Any]]]]
+    IpRangeWhitelist: Optional[List[Union[str, Dict[str, Any]]]]
+    SourceVpcBlacklist: Optional[List[Union[str, Dict[str, Any]]]]
+    SourceVpcWhitelist: Optional[List[Union[str, Dict[str, Any]]]]
+
+
+class CloudWatchEventProperties(BaseModel):
+    EventBusName: Optional[PassThrough]
+    Input: Optional[PassThrough]
+    InputPath: Optional[PassThrough]
+    Pattern: Optional[PassThrough]
+
+
+class CloudWatchEvent(BaseModel):
+    Type: Literal["CloudWatchEvent"]
+    Properties: CloudWatchEventProperties
+
+
+class EventBridgeRuleEventProperties(BaseModel):
+    DeadLetterConfig: Optional[DeadLetterConfig]
+    EventBusName: Optional[PassThrough]
+    Input: Optional[PassThrough]
+    InputPath: Optional[PassThrough]
+    Pattern: Optional[PassThrough]
+    RetryPolicy: Optional[PassThrough]
+
+
+class EventBridgeRuleEvent(BaseModel):
+    Type: Literal["EventBridgeRule"]
+    Properties: EventBridgeRuleEventProperties
+
+
+class Auth(BaseModel):
+    ApiKeyRequired: Optional[bool]
+    AuthorizationScopes: Optional[List[str]]
+    Authorizer: Optional[str]
+    ResourcePolicy: Optional[ResourcePolicy]
+
+
+class ApiEventProperties(BaseModel):
+    Auth: Optional[Auth]
+    Method: str
+    Path: str
+    RestApiId: Optional[SamIntrinsic]
+    UnescapeMappingTemplate: Optional[bool]  # TODO: Add to docs
+
+
+class ApiEvent(BaseModel):
+    Type: Literal["Api"]
+    Properties: ApiEventProperties
+
+
+class Properties(BaseModel):
+    Definition: Optional[Dict[str, Any]]
+    DefinitionSubstitutions: Optional[Dict[str, Any]]
+    DefinitionUri: Optional[Union[str, PassThrough]]
+    Events: Optional[
+        Dict[
+            str,
+            Union[
+                ScheduleEvent,
+                ScheduleV2Event,
+                CloudWatchEvent,
+                EventBridgeRuleEvent,
+                ApiEvent,
+            ],
+        ]
+    ]
+    Logging: Optional[PassThrough]
+    Name: Optional[PassThrough]
+    PermissionsBoundary: Optional[PassThrough]
+    Policies: Optional[Union[str, List[str], Dict[str, Any], List[Dict[str, Any]]]]
+    Role: Optional[PassThrough]
+    Tags: Optional[Dict[str, Any]]
+    Tracing: Optional[PassThrough]
+    Type: Optional[PassThrough]
+
+
+class Resource(BaseModel):
+    Type: Literal["AWS::Serverless::StateMachine"]
+    Properties: Properties
+    Condition: Optional[PassThrough]
